requirements.txt
azure-storage-blob>=12.20.0
azure-identity>=1.17.1
httpx>=0.27.2
python-dotenv>=1.0.1
pydantic>=2.8.2
tiktoken>=0.7.0
openai>=1.40.0
pytest>=8.3.2
mypy>=1.11.1
ruff>=0.5.7
